+++ example/lib/mobile_example.dart
@@ -5,40 +5,47 @@ import 'package:image_painter/image_painter.dart';
 import 'package:open_file/open_file.dart';
 import 'package:path_provider/path_provider.dart';
 
-class SignatureExample extends StatefulWidget {
+class MobileExample extends StatefulWidget {
+  const MobileExample({Key? key}) : super(key: key);
+
   @override
-  _SignatureExampleState createState() => _SignatureExampleState();
+  State<MobileExample> createState() => _MobileExampleState();
 }
 
-class _SignatureExampleState extends State<SignatureExample> {
-  final _controller = ImagePainterController();
+class _MobileExampleState extends State<MobileExample> {
+  final ImagePainterController _controller = ImagePainterController(
+    color: Colors.green,
+    strokeWidth: 4,
+    mode: PaintMode.line,
+  );
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text("Image Painter Example")),
-      body: Center(
-        child: ClipRRect(
-          borderRadius: BorderRadius.circular(8),
-          child: ImagePainter.signature(
-            height: 200,
-            width: 300,
-            controller: _controller,
-            signatureBgColor: Colors.grey[200],
-          ),
-        ),
+      appBar: AppBar(
+        title: const Text("Image Painter Example"),
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.save_alt),
+            onPressed: saveImage,
+          )
+        ],
       ),
-      floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.save),
-        onPressed: saveImage,
+      body: ImagePainter.asset(
+        "assets/sample.jpg",
+        controller: _controller,
+        scalable: true,
+        textDelegate: TextDelegate(),
       ),
     );
   }
 
   void saveImage() async {
     final image = await _controller.exportImage();
+    final imageName = '${DateTime.now().millisecondsSinceEpoch}.png';
     final directory = (await getApplicationDocumentsDirectory()).path;
     await Directory('$directory/sample').create(recursive: true);
-    final fullPath = '$directory/sample/image.png';
+    final fullPath = '$directory/sample/$imageName';
     final imgFile = File('$fullPath');
     if (image != null) {
       imgFile.writeAsBytesSync(image);
@@ -55,7 +62,9 @@ class _SignatureExampleState extends State<SignatureExample> {
                 onPressed: () => OpenFile.open("$fullPath"),
                 child: Text(
                   "Open",
-                  style: TextStyle(color: Colors.blue[200]),
+                  style: TextStyle(
+                    color: Colors.blue[200],
+                  ),
                 ),
               )
             ],