+++ lib/image_painter.dart
@@ -1,4 +1,5 @@
 library image_painter;
 
 export 'src/_paint_over_image.dart';
+export 'src/controller.dart';
 export 'src/delegates/text_delegate.dart';